chore: organize Next fidelity mirror adapters
