+++ AGENTS.md
@@ -120,7 +120,7 @@ Use Next's transform layer for source-level Next features instead of regex trans
 
 Keep compatibility shims small, explicit, and as close to Next.js as possible. If Next installs a global, injects a polyfill, or relies on webpack `ProvidePlugin` behavior, mirror that behavior through the narrowest Vite/Vitest adapter we can. Prefer importing Next's bootstrap/runtime code, or copying the relevant upstream block, over inventing a local substitute. Process, Buffer, WebSocket, document, and browser runtime shims are acceptable when they reflect real Next behavior; custom behavior beyond Next should be treated as high-risk, documented with a source reference, and covered by a regression test that fails without it.
 
-When copying code from Next.js, wrap it in clear copy markers and include the upstream source path or permalink plus a short adaptation note:
+When copying or adapting code from Next.js, wrap it in clear markers and include the upstream source path or permalink plus a short adaptation note. Use `Begin copy` / `End copy` only for mechanically copied upstream code. Use `Begin adapted` / `End adapted` for Vite/Vitest boundary code that deliberately translates concrete upstream Next loader, template, runtime, manifest plugin, compiler option, or routing conversion behavior:
 
 ```ts
 // Begin copy: Next.js <behavior/name>
@@ -128,6 +128,12 @@ When copying code from Next.js, wrap it in clear copy markers and include the up
 // Adaptation: <why this differs for Vite/Vitest tests>
 ...
 // End copy
+
+// Begin adapted: Next.js <behavior/name>
+// Source: https://github.com/vercel/next.js/blob/<sha>/<path>
+// Adaptation: <which Vite/Vitest boundary forces this translation>
+...
+// End adapted
 ```
 
 Tests should cover framework features, not just demo behavior. Every supported Next API, route convention, page export, or runtime behavior touched by the plugin should have a focused test in `playground/nextjs-notes-demo`. Package-level unit tests in `packages/vitest-plugin-rsc/src/nextjs` are still useful for plugin internals, transforms, aliases, and loader adapters, but they do not replace notes-demo coverage for user-visible Next behavior. Do not add app-local mocks to make the notes demo pass when the plugin can provide the behavior for every user.
