+++ packages/vitest-plugin-rsc/src/nextjs/app-render.test.ts
@@ -1,10 +1,10 @@
 import { Buffer } from "node:buffer";
 import { expect, test } from "vitest";
+import { createNextServerActionManifest } from "./src/build/webpack/plugins/flight-client-entry-plugin.ts";
 import {
-  createNextServerActionManifest,
   createViteRscClientModulesProxy,
   createViteRscModuleMappingProxy,
-} from "./app-render-manifest.ts";
+} from "./src/build/webpack/plugins/flight-manifest-plugin.ts";
 import { patchBufferIndexOfUint8ArrayNeedle } from "./buffer-compat.ts";
 
 const patchedBufferIndexOfSymbol = Symbol.for("vitest-plugin-rsc.nextjs.patchedBufferIndexOf");
